Fix re-register error caused by stale framework ID, and stopped webserver from blocking after fatal framework error.
Added the error hook to the calico scheduler, and check for error
messages to see if registration was prevented due to a bad framework ID.
If so, wipe /calico in zk.

Also switched webserver from subprocess to multiprocess, as there's no
easy way to kill subprocess threads, and they will block a program from
exiting until their process is done. (And it near impossible to kill a
flask app with anything besides a sigint).